Price changes

+++ src/main/java/cart/Cart.java
@@ -2,10 +2,12 @@ package cart;
 
 import products.BaseProduct;
 
+import java.math.BigDecimal;
+
 public interface Cart {
     void add(BaseProduct product);
 
     void remove(BaseProduct product);
 
-    Double getSum();
+    BigDecimal getSum();
 }
